ci/job_generator/requirements.txt
allpairspy == 2.5.0
typeguard
pyaml
types-PyYAML
